Fixed up homepairs hamburger menu and fixed alignment issue in non-Chrome browsers
