--- homepairs/homepairsUI/res/__mocks__images.tsx
@@ -1,33 +1,0 @@
-const blender = null;
-const bolt = null;
-const calendar = null;
-const clock = null;
-const defaultProperty = null;
-const downArrow = null;
-const fan = null;
-const hamburger = null;
-const icon = null;
-const roopairsLogo = null;
-const splash = null;
-const testLogo = null;
-const tint = null;
-const trash = null;
-const upArrow = null;
-
-export { 
-	blender,
-	bolt,
-	calendar,
-	clock,
-	defaultProperty,
-	downArrow,
-	fan,
-	hamburger,
-	icon,
-	roopairsLogo,
-	splash,
-	testLogo,
-	tint,
-	trash,
-	upArrow
-}
