+++ homepairs/homepairsUI/src/Elements/index.tsx
@@ -22,29 +22,20 @@
  */
 
 export * from './Forms/InputForm';
-export * from './Buttons/ThinButton';
 export * from './Panels/AppliancePanel';
 export * from './Panels/ApplianceCategoryPanel';
-export * from './Buttons/ServiceRequestButton';
-export * from './Buttons/ServiceProviderButton';
 export * from './Cards/Card';
 export * from './Stickers/Sticker';
 export * from './Tiles/Tiles';
-
+export * from './Buttons/index';
 export {default as DatePicker} from './DatePicker/DatePicker';
-export {default as ButtonWithBitmap} from './Buttons/ButtonWithBitmap';
 export {default as InputForm } from './Forms/InputForm';
 export {default as GoogleInputForm} from './Forms/GoogleInputForm';
-export {default as ThinButton} from './Buttons/ThinButton';
-export {default as LoginButton} from './Buttons/LoginButton';
 export {default as Card } from './Cards/Card';
 export {default as Sticker} from './Stickers/Sticker';
-export {default as HamburgerButton} from './Buttons/HamburgerButton';
 export {default as AppliancePanel} from './Panels/AppliancePanel';
 export {default as ServiceRequestAddressPanel} from './Panels/ServiceRequestAddressPanel';
 export {default as AddressPanel} from './Panels/AddressPanel';
 export {default as ApplianceCategoryPanel} from './Panels/ApplianceCategoryPanel';
 export {default as ServiceTypePanel} from './Panels/ServiceTypePanel';
-export {default as ServiceRequestButton} from './Buttons/ServiceRequestButton';
-export {default as ServiceProviderButton} from './Buttons/ServiceProviderButton';
 export {default as LocationItem } from './LocationItem/LocationItem';
